+++ json/sonic/doc.go
@@ -1,0 +1,2 @@
+// Package sonic 提供基于 github.com/bytedance/sonic 的 Arkarta JSON 高性能实现。
+package sonic